TutorialScreen.kt, day: Long

+++ app/src/main/java/com/example/goneappforandroid/TasksViewModel.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.launch
 
 class TasksViewModel(private val repository: TasksRepository): ViewModel() {
 
-    fun insertTask(text: String, minute: Int, hour: Int, day: Int, checked: Boolean){
+    fun insertTask(text: String, minute: Int, hour: Int, day: Long, checked: Boolean){
         viewModelScope.launch(Dispatchers.IO) {
             repository.insertTask(Task(0, text, minute, hour, day, checked))
         }
